requirements_dev.txt
-r requirements.txt

bpython==0.18
django-debug-toolbar==2.0
django-extensions==2.2.1
